feat: voter projet + formulaire projet corrigé + sécurité granulaire

+++ src/Controller/ProjectController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Controller;
 
 use App\Entity\Project;
@@ -31,6 +30,7 @@ final class ProjectController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $project->setUser($this->getUser());
+            $project->setCreatedAt(new \DateTimeImmutable());
             $entityManager->persist($project);
             $entityManager->flush();
 
@@ -46,6 +46,8 @@ final class ProjectController extends AbstractController
     #[Route('/{id}', name: 'app_project_show', methods: ['GET'])]
     public function show(Project $project): Response
     {
+        $this->denyAccessUnlessGranted('PROJECT_VIEW', $project);
+
         return $this->render('project/show.html.twig', [
             'project' => $project,
         ]);
@@ -54,6 +56,8 @@ final class ProjectController extends AbstractController
     #[Route('/{id}/edit', name: 'app_project_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Project $project, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('PROJECT_EDIT', $project);
+
         $form = $this->createForm(ProjectType::class, $project);
         $form->handleRequest($request);
 
@@ -72,6 +76,8 @@ final class ProjectController extends AbstractController
     #[Route('/{id}', name: 'app_project_delete', methods: ['POST'])]
     public function delete(Request $request, Project $project, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('PROJECT_DELETE', $project);
+
         if ($this->isCsrfTokenValid('delete'.$project->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($project);
             $entityManager->flush();
